requirements.txt
anyio==3.3.4
asgiref==3.2.10
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.1.1
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.10.0
django-oidc-provider==0.7.0
djangorestframework==3.12.4
drf-oidc-auth==0.9
future==0.18.2
h11==0.12.0
httpcore==0.13.7
httpx==0.20.0
idna==3.3
mysqlclient==2.0.3
onesignal-sdk==1.1.0
pycryptodomex==3.11.0
pyjwkest==1.4.2
python-dateutil==2.8.2
python-dotenv==0.19.1
pytz==2021.3
requests==2.26.0
rfc3986==1.5.0
six==1.16.0
sniffio==1.2.0
sqlparse==0.4.2
urllib3==1.26.7
